Seg fault on removing objects (to fix)

+++ src/Objects/Object.cpp
@@ -67,7 +67,7 @@ namespace pheng {
 
     }
 
-    double Object::calculateEnergy(double h, double g) {
+    float Object::calculateEnergy(double h, double g) {
         return 0.5 * Mass * Velocity.norm() * Velocity.norm() +
                Mass * g * (h - Position.getY());
     }
